changed: tests re-implemented using CUnit, by stylin
r3066

+++ FreeBASIC/tests/dim/submod/submod_b.bas
@@ -1,7 +1,9 @@
 
+
+
 #include once "submod_c.bi"
 
-	dim shared ext_statarray(1 to 21, 1 to 2) as integer
+	dim shared ext_statarray(1 to 21, 1 to 2)
 
 	redim ext_dynarray(1 to 21, 1 to 2)
